WEP-50 move app related files to /app
